+++ src/App/current_list_actions.js
@@ -1,11 +1,12 @@
 // @flow
 import pushid from "pushid";
 
-import type { List } from "@src/types";
 import { type Action, makeAction, modifyState } from "@src/store";
-import type { AppState } from "@src/state/app";
+import * as Types from "@src/types";
 
-const modifyCurrentList = (fn: List => List) => (state: AppState): AppState => {
+const modifyCurrentList = (fn: Types.List => Types.List) => (
+  state: Types.AppState
+): Types.AppState => {
   const currentListId = state.navigationState.listId;
   if (!currentListId) {
     throw new Error("no current list");
@@ -16,7 +17,7 @@ const modifyCurrentList = (fn: List => List) => (state: AppState): AppState => {
   };
 };
 
-export const selectCurrentList = (state: AppState): ?List => {
+export const _selectCurrentList = (state: Types.AppState): ?Types.List => {
   if (state.navigationState.listId == null) {
     return;
   }
@@ -24,7 +25,7 @@ export const selectCurrentList = (state: AppState): ?List => {
   return state.lists[currentListId];
 };
 
-const _addTodo = () => (list: List): List => {
+const _addTodo = () => (list: Types.List): Types.List => {
   const newTodo = {
     content: "",
     state: "in-progress"
@@ -34,7 +35,7 @@ const _addTodo = () => (list: List): List => {
   };
 };
 
-const _deleteTodo = (todoId: string) => (list: List): List => {
+const _deleteTodo = (todoId: string) => (list: Types.List): Types.List => {
   const newTodos = { ...list.todos };
   delete newTodos[todoId];
   return {
@@ -43,8 +44,8 @@ const _deleteTodo = (todoId: string) => (list: List): List => {
 };
 
 const _setTodoContent = (todoId: string, content: string) => (
-  list: List
-): List => {
+  list: Types.List
+): Types.List => {
   const todo = list.todos[todoId];
   if (!todo) {
     throw new Error("todoId not found");
