--- src/state/app.js
@@ -1,22 +1,0 @@
-// @flow
-import type { List } from "@src/types";
-
-export type AppState = {|
-  lists: {
-    [listId: string]: List
-  },
-  navigationState: {|
-    listId: void | string
-  |}
-|};
-
-export const defaultState: AppState = {
-  lists: {
-    list1: {
-      todos: {}
-    }
-  },
-  navigationState: {
-    listId: "list1"
-  }
-};
