+++ src/storage.js
@@ -1,11 +1,11 @@
 // @flow
-import type { AppState } from "@src/state/app";
+import * as Types from "@src/types";
 
-export function saveToLocalStorage(state: AppState): void {
+export function saveToLocalStorage(state: Types.AppState): void {
   global.localStorage.setItem("App", JSON.stringify(state));
 }
 
-export function loadFromLocalStorage(): ?AppState {
+export function loadFromLocalStorage(): ?Types.AppState {
   const item = global.localStorage.getItem("App");
   if (!item) {
     return;
