+++ src/types.js
@@ -9,3 +9,23 @@ export type Todo = {
 export type List = {
   todos: { [string]: Todo }
 };
+
+export type AppState = {|
+  lists: {
+    [listId: string]: List
+  },
+  navigationState: {|
+    listId: void | string
+  |}
+|};
+
+export const defaultAppState: AppState = {
+  lists: {
+    list1: {
+      todos: {}
+    }
+  },
+  navigationState: {
+    listId: "list1"
+  }
+};
